+++ fuse/plugins/micro_physics/cuts_and_selections/__init__.py
@@ -1,6 +1,3 @@
-from . import apply_selections
-from .apply_selections import *
-
 from . import detector_volumes
 from .detector_volumes import *
